fix: Reorganize code in directories and fix import paths
