+++ experiments/sb3_custom_comparison/README.md
@@ -1,0 +1,11 @@
+THis is an experiment comparing results of 2 PPO implementations: my custom PPO at `src/algorithms/PPO_algorithm.py` with stable-baselines-3 implementation.
+
+At the moment of experiment, source code of custom ppo looks as follows:
+
+```py
+Add code here
+```
+
+The comparison script can be found at `experiments/sb3_custom_comparison/ppo_comparison.py`:
+
+Results of the comparison can be found at
